feat!: add generated, font-based icon themes
implement a system that generates file icon themes from a single set of
svg icons. this system will use fantasticon to generate a the icon font
file and other assets for consuming the icons.

marked as a breaking change because it deprecates the old svg icon
themes (despite the general backwards compatibility of the new system).

this one's a big one, but it can't be helped :sob: